♻️ [Refactor] processDifferences 함수 분리
processDifferences 함수 분리
